First outline created of version-1.md
Added initial outline of what the to-be was going to be and some screenshots of the pages and flow from Mural board. Still need to add alt text to most images and captions to help focus what has changed or the main points to be picked up. 

Will look at if flow diagram functionality within GitHub will work or be useful.