Added image alt text and started captioning
Captions could use more detail and better detail of what we think is necessary to explain as part of this document. Linking to Mural could be a better idea to get across all the content - or a Google doc might be more accessible and shareable for future iterations of error content.