Update version-1.md
Made some small changes for readability and house style. 

I have a few queries but didn't want to put these within the text itself, so I'll try adding these once I've opened a PR request. 

Would be good to discuss the best way to do this as I'm not sure!