Update README.md
Have updated wording following discussion with Hannah.

New wording: However, this means form fillers have to decide whether they think a question is relevant to them, and they may decide not to answer when in fact they need to. It also means that form creators may feel the need to clarify who should answer the question in the 'hint text', which may or may not be read by the form completer.

Old wording: However, this means form creators need to give questions context within the “hint text” to help form completers decide if they need to complete a question or not.