Update auth diagram with session timeout
Added the missing redirect to forms admin/auth/auth0.

I've split Auth0 into two actors to reflect the different roles it's
playing in our system.
Originally this was to show the difference between our login session and the [auth0
session](https://auth0.com/docs/manage-users/sessions/session-layers).

[It turns out this doesn't need to be documented
yet](https://trello.com/c/rmk3P0kO/1299-fix-auth0-single-sign-on), but I
think it still might be useful. It would be good to compare with the
google workspace diagram.

It could be that auth0 is outside our control so we should treat
it as a black box.

I've changed the references to session to be more precise about what's
happening. Talking in terms of the cookie values is lower level but
gives [this
ticket](https://trello.com/c/yi3v9Mgo/1121-require-forms-admin-users-to-re-authenticate-after-24-hours)
a clearer description.

Session is the right term but it is also implies more is happening
behind the scenes, which I don't think there is here.