Remove class-diagram for forms-api
The forms-api app is no longer used in any part of our system, its data
and responsibilities have been moved to the forms-admin app, and this
class diagram is no longer relevant and can be removed.

All of the entity types that were in the forms-api class diagram are now
also in the forms-admin class diagram, so nothing should be lost by
this.