Remove forms-api from sequence diagrams
The forms-api app is no longer used in any part of our system, its data
and responsibilities have been moved to the forms-admin app, and
forms-runner retrieves form definitions from forms-admin.

The sequence diagrams in this repo need to be updated to reflect that,
this commit does so.

This does mean that there is less detail in some sequence diagrams,
however this is accurate as to our external API surface; forms-admin
exposes fewer public API endpoints than forms-api did.